Use StructuredSerialize/StructuredDeserialize (#171)
Minimal changes to replace uses of StructuredClone with StructuredSerialize/StructuredDeserialize, per https://github.com/w3c/IndexedDB/issues/170 to track latest [HTML] definitions.

Note that when put() (etc) is called that this does not simply serialize then operate on the output record - the record is immediately deserialized to a clone in an abstract targetRealm so that subsequent operations (extract a key, inject a key, etc) are not updated. I added non-normative details explaining this at the various sites.

(Somewhat coincidentally, that's how Blink is actually implemented, since the output of serialization is a opaque set of bytes rather than an easily-traversed structure.)

The definition of operations that operate on the clone could be changed to operate on the record instead. This would require redoing the "extract a key from a value" and "inject a key into a value" section (replace "value" with "record"), and duplicating the "convert a key to a value" and "convert a value to a key" steps with "record" variants.