Editorial: Fix _includes link, add note.
"includes" is reserved word in WebIDL, so a method with that name
needs to be escaped with a leading underscore in IDL fragments.
This confuses Bikeshed (tabatkins/bikeshed#1489) so manually
correct the link. Also, the 2.0 version of the IDB spec had a note
which gave context for this; transplant that to the 3.0 version.

Resolves #270