Fix links to "Type" in ECMAScript
Bikeshed converted `[=/Type=]` to a link to a definition in the WebCrypto API,
which was not the intended target. This got fix in the cross-reference database
used by Bikeshed, but that also means that `[=/Type=]` no longer links to
anything.

This fixes these links to `[=ECMAScript/Type=]` (defined in the anchors metadata
at the top of the source).