Drop now useless ECMAScript anchor definitions
Most local anchor definitions that targeted the ECMAScript spec are no longer
needed because Bikeshed now has the spec in its cross-references database.
This update replaces them with the right shorthand text.

Remaining anchor definitions target sections or concepts that are not real
definitions and as such are not in the cross-references database.