Adjust links to ECMAScript specifications (#401)
* Fix links to "Type" in ECMAScript

Bikeshed converted `[=/Type=]` to a link to a definition in the WebCrypto API,
which was not the intended target. This got fix in the cross-reference database
used by Bikeshed, but that also means that `[=/Type=]` no longer links to
anything.

This fixes these links to `[=ECMAScript/Type=]` (defined in the anchors metadata
at the top of the source).

* Drop now useless ECMAScript anchor definitions

Most local anchor definitions that targeted the ECMAScript spec are no longer
needed because Bikeshed now has the spec in its cross-references database.
This update replaces them with the right shorthand text.

Remaining anchor definitions target sections or concepts that are not real
definitions and as such are not in the cross-references database.